refactor: update import paths and consolidate error message component
