+++ lib/core/shared/common_ui/components/auth_header.dart
