+++ lib/core/shared/shared_services/heptics/heptic_manager.dart
