+++ lib/core/utils/message_service.dart
@@ -1,4 +1,4 @@
-import 'package:defifundr_mobile/core/shared/user_interface/overlay/overlay_message.dart';
+import 'package:defifundr_mobile/core/shared/common_ui/overlay/overlay_message.dart';
 import 'package:defifundr_mobile/core/utils/enum_file.dart';
 import 'package:flutter/material.dart';
