+++ lib/modules/authentication/login/presentation/login_screen/login_screen.dart
@@ -6,10 +6,10 @@ import 'package:defifundr_mobile/core/design_system/app_colors/app_colors.dart';
 import 'package:defifundr_mobile/core/design_system/theme_extension/app_theme_extension.dart';
 import 'package:defifundr_mobile/core/enums/app_text_field_enums.dart';
 import 'package:defifundr_mobile/core/routers/routes_constant.dart';
-import 'package:defifundr_mobile/core/shared/user_interface/appbar/appbar_header.dart';
-import 'package:defifundr_mobile/core/shared/user_interface/components/or_widget.dart';
-import 'package:defifundr_mobile/core/shared/user_interface/textfield/app_text_field.dart';
-import 'package:defifundr_mobile/core/shared/user_interface/buttons/primary_button.dart';
+import 'package:defifundr_mobile/core/shared/common_ui/appbar/appbar_header.dart';
+import 'package:defifundr_mobile/core/shared/common_ui/components/or_widget.dart';
+import 'package:defifundr_mobile/core/shared/common_ui/textfield/app_text_field.dart';
+import 'package:defifundr_mobile/core/shared/common_ui/buttons/primary_button.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
