+++ lib/modules/finance/presentation/address_book_screen.dart
@@ -6,7 +6,7 @@ import 'package:go_router/go_router.dart';
 import 'package:defifundr_mobile/core/routers/routes_constant.dart';
 import 'package:defifundr_mobile/modules/finance/presentation/add_address_screen.dart';
 import 'dart:math';
-import 'package:defifundr_mobile/core/shared/user_interface/appbar/appbar.dart'; // Import DeFiRaiseAppBar
+import 'package:defifundr_mobile/core/shared/common_ui/appbar/appbar.dart'; // Import DeFiRaiseAppBar
 
 // Define a data model for a saved address entry
 class SavedAddress {
