+++ lib/modules/onboarding/presentation/individual_account_flow/screens/address_details_screen.dart
@@ -3,7 +3,7 @@
 import 'package:defifundr_mobile/core/design_system/theme_extension/app_theme_extension.dart';
 import 'package:defifundr_mobile/core/routers/routes_constant.dart';
 import 'package:flutter/material.dart';
-import 'package:defifundr_mobile/core/shared/user_interface/textfield/app_text_field.dart';
+import 'package:defifundr_mobile/core/shared/common_ui/textfield/app_text_field.dart';
 
 import 'package:go_router/go_router.dart';
