+++ lib/modules/onboarding/presentation/multi_factor_authentication_screen/screens/enable_fingerprint_screen.dart
@@ -1,8 +1,8 @@
 import 'package:defifundr_mobile/core/design_system/theme_extension/app_theme_extension.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
-import '../../../../../core/shared/user_interface/buttons/primary_button.dart';
-import '../../../../../core/shared/user_interface/buttons/secondary_button.dart';
+import '../../../../../core/shared/common_ui/buttons/primary_button.dart';
+import '../../../../../core/shared/common_ui/buttons/secondary_button.dart';
 
 class EnableFingerprintScreen extends StatelessWidget {
   const EnableFingerprintScreen({super.key});
